Correção de estrutura de projeto final e remoção de comentários
