+++ SkinArt_Manager/Controllers/UsuarioController.cs
@@ -4,8 +4,6 @@ using SkinArt_Manager.DTOs.UsuarioDTO;
 using SkinArt_Manager.Models;
 using SkinArt_Manager.Services;
 
-// ----> Comentário
-
 namespace SkinArt_Manager.Controllers
 {
     [ApiController]
