+++ SkinArt_Manager/DTOs/ClienteDTO/AtualizarClienteDTO.cs
@@ -1,10 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace SkinArt_Manager.DTOs
+namespace SkinArt_Manager.DTOs.ClienteDTO
 {
-    // ----> Comentário
-    public class CriarClienteDTO
+    public class AtualizarClienteDTO
     {
+        [Required(ErrorMessage = "O ID do cliente é obrigatório.")]
+        public int Id { get; set; }
+
         [Required(ErrorMessage = "O nome completo é obrigatório.")]
         [MaxLength(255, ErrorMessage = "O nome completo não pode exceder 255 caracteres.")]
         public string NomeCompleto { get; set; } = string.Empty;
