+++ SkinArt_Manager/DTOs/EstoqueDTO/CriarEstoqueItemDTO.cs
@@ -1,10 +1,7 @@
-﻿namespace SkinArt_Manager.DTOs
+﻿namespace SkinArt_Manager.DTOs.EstoqueDTO
 {
-    // ----> Comentário
-
-    public class AtualizarEstoqueItemDTO
+    public class CriarEstoqueItemDTO
     {
-        public int Id { get; set; }
         public string Categoria { get; set; }
         public string NomeItem { get; set; }
         public int Quantidade { get; set; }
