+++ SkinArt_Manager/DTOs/EstoqueDTO/EstoqueItemDTO.cs
@@ -1,8 +1,7 @@
-﻿namespace SkinArt_Manager.DTOs
+﻿namespace SkinArt_Manager.DTOs.EstoqueDTO
 {
     public class EstoqueItemDTO
     {
-        // ----> Comentário
         public int Id { get; set; }
         public string Categoria { get; set; }
         public string NomeItem { get; set; }
