+++ SkinArt_Manager/Services/TokenService.cs
@@ -6,7 +6,6 @@ using System.Text;
 
 namespace SkinArt_Manager.Services
 {
-    // ----> Comentário
     public class TokenService
     {
         public static string GenerateToken(LoginResponseDTO usuario)
